# Store dense_vector values as bfloat16