--- a/x-pack/plugin/src/test/resources/rest-api-spec/test/vectors/10_dense_vector_basic.yml
+++ b/x-pack/plugin/src/test/resources/rest-api-spec/test/vectors/10_dense_vector_basic.yml
@@ -58,17 +58,20 @@ setup:
 
   - match: {hits.total: 3}
 
+  # result should be 65425.624, but we allow 1% error because vectors are indexed as bfloat16
   - match: {hits.hits.0._id: "1"}
-  - gte: {hits.hits.0._score: 65425.62}
-  - lte: {hits.hits.0._score: 65425.63}
+  - gte: {hits.hits.0._score: 62154.34}
+  - lte: {hits.hits.0._score: 68696.91}
 
+  # result should be 37111.98, but we allow 1% error because vectors are indexed as bfloat16
   - match: {hits.hits.1._id: "3"}
-  - gte: {hits.hits.1._score: 37111.98}
-  - lte: {hits.hits.1._score: 37111.99}
+  - gte: {hits.hits.1._score: 35256.38}
+  - lte: {hits.hits.1._score: 38967.58}
 
+  # result should be 35853.78, but we allow 1% error because vectors are indexed as bfloat16
   - match: {hits.hits.2._id: "2"}
-  - gte: {hits.hits.2._score: 35853.78}
-  - lte: {hits.hits.2._score: 35853.79}
+  - gte: {hits.hits.2._score: 34061.09}
+  - lte: {hits.hits.2._score: 37646.47}
 
 ---
 "Cosine Similarity":