--- a/x-pack/plugin/vectors/src/main/java/org/elasticsearch/xpack/vectors/mapper/VectorEncoderDecoder.java
+++ b/x-pack/plugin/vectors/src/main/java/org/elasticsearch/xpack/vectors/mapper/VectorEncoderDecoder.java
@@ -158,7 +158,7 @@ public final class VectorEncoderDecoder {
 
     public static int denseVectorLength(Version indexVersion, BytesRef vectorBR) {
         return indexVersion.onOrAfter(Version.V_7_5_0)
-            ? (vectorBR.length - INT_BYTES) / INT_BYTES
+            ? (vectorBR.length - INT_BYTES) / SHORT_BYTES
             : vectorBR.length / INT_BYTES;
     }