Currently dense_vector values are stored as usual floats of 32 bits.
This patch make vector values to be stored as bfloat16,
which should twice reduce index size for dense_vector field
at the expense of loss of some precision (less that 1% relative
error).

Closes #46980