@mayya-sharipova 2019-09-24
Benchmarks results
5 shards ES node, 5Gb heap

bfloat16:
- produced always 100% recall
- has index size less that the current master
- performance for queries per second usually better than current master, except a case with `glove-100-angular` where performance was slightly worse

| dataset | master qps | bfloat16  qps| master index size| bfloat16 index size |
| :---  |---: | ---: | ---: | ---: | 
|sift-128-euclidean | 13.187 | 13.223| 840.9mb |  596.8mb| 
| glove-100-angular | 15.287 | **14.864**| 2.1gb|  1.9gb|
| stackoverflow-512-angular | 3.267 | 4.848 | 9.5gb | 8.5gb|