@mayya-sharipova 2020-03-02
Closing this, as we have decided not to implement bfloat16 for now.